requirements.txt
asgiref==3.2.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-auto-prefetching==0.1.7
django-cors-headers==3.2.1
django-filter==2.2.0
django-silk==4.0.0
djangorestframework==3.11.0
gprof2dot==2019.11.30
idna==2.8
Jinja2==2.11.1
MarkupSafe==1.1.1
pycodestyle==2.5.0
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
